+++ src/appshell/internal/applicationactioncontroller.cpp
@@ -23,6 +23,7 @@
 
 #include <QCoreApplication>
 #include <QCloseEvent>
+#include <QWindow>
 
 #include "translation.h"
