+++ src/engraving/libmscore/layout/layout.cpp
@@ -42,8 +42,8 @@
 #include "../stafflines.h"
 #include "../tuplet.h"
 #include "../tie.h"
-#include "../layout.h"
 
+#include "layoutcontext.h"
 #include "layoutmeasure.h"
 
 using namespace mu::engraving;
@@ -69,175 +69,6 @@ static bool inline almostZero(qreal value)
 {
     // 1e-3 is close enough to zero to see it as zero.
     return value > -1e-3 && value < 1e-3;
-}
-
-//---------------------------------------------------------
-// validMMRestMeasure
-//    return true if this might be a measure in a
-//    multi measure rest
-//---------------------------------------------------------
-
-static bool validMMRestMeasure(Measure* m)
-{
-    if (m->irregular()) {
-        return false;
-    }
-
-    int n = 0;
-    for (Segment* s = m->first(); s; s = s->next()) {
-        for (Element* e : s->annotations()) {
-            if (!(e->isRehearsalMark() || e->isTempoText() || e->isHarmony() || e->isStaffText() || e->isSystemText()
-                  || e->isInstrumentChange())) {
-                return false;
-            }
-        }
-        if (s->isChordRestType()) {
-            bool restFound = false;
-            int tracks = m->score()->ntracks();
-            for (int track = 0; track < tracks; ++track) {
-                if ((track % VOICES) == 0 && !m->score()->staff(track / VOICES)->show()) {
-                    track += VOICES - 1;
-                    continue;
-                }
-                if (s->element(track)) {
-                    if (!s->element(track)->isRest()) {
-                        return false;
-                    }
-                    restFound = true;
-                }
-            }
-            for (Element* e : s->annotations()) {
-                if (e->isFermata()) {
-                    return false;
-                }
-            }
-            if (restFound) {
-                ++n;
-            }
-            // measure is not empty if there is more than one rest
-            if (n > 1) {
-                return false;
-            }
-        }
-    }
-    return true;
-}
-
-//---------------------------------------------------------
-//  breakMultiMeasureRest
-//    return true if this measure should start a new
-//    multi measure rest
-//---------------------------------------------------------
-
-static bool breakMultiMeasureRest(Measure* m)
-{
-    if (m->breakMultiMeasureRest()) {
-        return true;
-    }
-
-    if (m->repeatStart()
-        || (m->prevMeasure() && m->prevMeasure()->repeatEnd())
-        || (m->isIrregular())
-        || (m->prevMeasure() && m->prevMeasure()->isIrregular())
-        || (m->prevMeasure() && (m->prevMeasure()->sectionBreak()))) {
-        return true;
-    }
-
-    auto sl = m->score()->spannerMap().findOverlapping(m->tick().ticks(), m->endTick().ticks());
-    for (auto i : sl) {
-        Spanner* s = i.value;
-        // break for first measure of volta or textline and first measure *after* volta
-        if ((s->isVolta() || s->isTextLine()) && (s->tick() == m->tick() || s->tick2() == m->tick())) {
-            return true;
-        }
-    }
-
-    // break for marker in this measure
-    for (Element* e : m->el()) {
-        if (e->isMarker()) {
-            Marker* mark = toMarker(e);
-            if (!(mark->align() & Align::RIGHT)) {
-                return true;
-            }
-        }
-    }
-
-    // break for marker & jump in previous measure
-    Measure* pm = m->prevMeasure();
-    if (pm) {
-        for (Element* e : pm->el()) {
-            if (e->isJump()) {
-                return true;
-            } else if (e->isMarker()) {
-                Marker* mark = toMarker(e);
-                if (mark->align() & Align::RIGHT) {
-                    return true;
-                }
-            }
-        }
-    }
-
-    // break for MeasureRepeat group
-    for (int staffIdx = 0; staffIdx < m->score()->nstaves(); ++staffIdx) {
-        if (m->isMeasureRepeatGroup(staffIdx)
-            || (m->prevMeasure() && m->prevMeasure()->isMeasureRepeatGroup(staffIdx))) {
-            return true;
-        }
-    }
-
-    for (Segment* s = m->first(); s; s = s->next()) {
-        for (Element* e : s->annotations()) {
-            if (!e->visible()) {
-                continue;
-            }
-            if (e->isRehearsalMark()
-                || e->isTempoText()
-                || ((e->isHarmony() || e->isStaffText() || e->isSystemText() || e->isInstrumentChange())
-                    && (e->systemFlag() || m->score()->staff(e->staffIdx())->show()))) {
-                return true;
-            }
-        }
-        for (int staffIdx = 0; staffIdx < m->score()->nstaves(); ++staffIdx) {
-            if (!m->score()->staff(staffIdx)->show()) {
-                continue;
-            }
-            Element* e = s->element(staffIdx * VOICES);
-            if (!e || e->generated()) {
-                continue;
-            }
-            if (s->isStartRepeatBarLineType()) {
-                return true;
-            }
-            if (s->isType(SegmentType::KeySig | SegmentType::TimeSig) && m->tick().isNotZero()) {
-                return true;
-            }
-            if (s->isClefType()) {
-                if (s->tick() != m->endTick() && m->tick().isNotZero()) {
-                    return true;
-                }
-            }
-        }
-    }
-    if (pm) {
-        Segment* s = pm->findSegmentR(SegmentType::EndBarLine, pm->ticks());
-        if (s) {
-            for (int staffIdx = 0; staffIdx < s->score()->nstaves(); ++staffIdx) {
-                BarLine* bl = toBarLine(s->element(staffIdx * VOICES));
-                if (bl) {
-                    BarLineType t = bl->barLineType();
-                    if (t != BarLineType::NORMAL && t != BarLineType::BROKEN && t != BarLineType::DOTTED && !bl->generated()) {
-                        return true;
-                    } else {
-                        break;
-                    }
-                }
-            }
-        }
-        if (pm->findSegment(SegmentType::Clef, m->tick())) {
-            return true;
-        }
-    }
-    return false;
 }
 
 Layout::Layout(Ms::Score* score)
@@ -567,28 +398,6 @@ void Layout::collectLinearSystem(LayoutContext& lc)
 }
 
 //---------------------------------------------------------
-//   layoutDrumsetChord
-//---------------------------------------------------------
-
-static void layoutDrumsetChord(Chord* c, const Drumset* drumset, const StaffType* st, qreal spatium)
-{
-    for (Note* note : c->notes()) {
-        int pitch = note->pitch();
-        if (!drumset->isValid(pitch)) {
-            // qDebug("unmapped drum note %d", pitch);
-        } else if (!note->fixed()) {
-            note->undoChangeProperty(Pid::HEAD_GROUP, int(drumset->noteHead(pitch)));
-            int line = drumset->line(pitch);
-            note->setLine(line);
-
-            int off  = st->stepOffset();
-            qreal ld = st->lineDistance().val();
-            note->rypos()  = (line + off * 2.0) * spatium * .5 * ld;
-        }
-    }
-}
-
-//---------------------------------------------------------
 //   extendedStemLenWithTwoNotesTremolo
 //    Goal: To extend stem of one of the chords to make the tremolo less steep
 //    Returns a modified pair of stem lengths of two chords
@@ -641,299 +450,4 @@ std::pair<qreal, qreal> Layout::extendedStemLenWithTwoNoteTremolo(Tremolo* tremo
 #endif
 
     return { stemLen1, stemLen2 };
-}
-
-//---------------------------------------------------------
-//   processLines
-//---------------------------------------------------------
-
-static void processLines(System* system, std::vector<Spanner*> lines, bool align)
-{
-    std::vector<SpannerSegment*> segments;
-    for (Spanner* sp : lines) {
-        SpannerSegment* ss = sp->layoutSystem(system);         // create/layout spanner segment for this system
-        if (ss->autoplace()) {
-            segments.push_back(ss);
-        }
-    }
-
-    if (align && segments.size() > 1) {
-        const int nstaves = system->staves()->size();
-        constexpr qreal minY = -1000000.0;
-        const qreal defaultY = segments[0]->rypos();
-        std::vector<qreal> y(nstaves, minY);
-
-        for (SpannerSegment* ss : segments) {
-            if (ss->visible()) {
-                qreal& staffY = y[ss->staffIdx()];
-                staffY = qMax(staffY, ss->rypos());
-            }
-        }
-        for (SpannerSegment* ss : segments) {
-            if (!ss->isStyled(Pid::OFFSET)) {
-                continue;
-            }
-            const qreal staffY = y[ss->staffIdx()];
-            if (staffY > minY) {
-                ss->rypos() = staffY;
-            } else {
-                ss->rypos() = defaultY;
-            }
-        }
-    }
-
-    //
-    // add shapes to skyline
-    //
-    for (SpannerSegment* ss : segments) {
-        if (ss->addToSkyline()) {
-            system->staff(ss->staffIdx())->skyline().add(ss->shape().translated(ss->pos()));
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   isTopTuplet
-//    returns true for the first CR of a tuplet that is not cross-staff
-//---------------------------------------------------------
-
-static bool isTopTuplet(ChordRest* cr)
-{
-    Tuplet* t = cr->tuplet();
-    if (t && t->elements().front() == cr) {
-        // find top level tuplet
-        while (t->tuplet()) {
-            t = t->tuplet();
-        }
-        // consider tuplet cross if anything moved within it
-        if (t->cross()) {
-            return false;
-        } else {
-            return true;
-        }
-    }
-
-    // no tuplet or not first element
-    return false;
-}
-
-//---------------------------------------------------------
-//   layoutTies
-//---------------------------------------------------------
-
-static void layoutTies(Chord* ch, System* system, const Fraction& stick)
-{
-    SysStaff* staff = system->staff(ch->staffIdx());
-    if (!staff->show()) {
-        return;
-    }
-    for (Note* note : ch->notes()) {
-        Tie* t = note->tieFor();
-        if (t) {
-            TieSegment* ts = t->layoutFor(system);
-            if (ts && ts->addToSkyline()) {
-                staff->skyline().add(ts->shape().translated(ts->pos()));
-            }
-        }
-        t = note->tieBack();
-        if (t) {
-            if (t->startNote()->tick() < stick) {
-                TieSegment* ts = t->layoutBack(system);
-                if (ts && ts->addToSkyline()) {
-                    staff->skyline().add(ts->shape().translated(ts->pos()));
-                }
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   layoutHarmonies
-//---------------------------------------------------------
-
-static void layoutHarmonies(const std::vector<Segment*>& sl)
-{
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if (e->isHarmony()) {
-                Harmony* h = toHarmony(e);
-                // For chord symbols that coincide with a chord or rest,
-                // a partial layout can also happen (if needed) during ChordRest layout
-                // in order to calculate a bbox and allocate its shape to the ChordRest.
-                // But that layout (if it happens at all) does not do autoplace,
-                // so we need the full layout here.
-                h->layout();
-                h->autoplaceSegmentElement();
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   alignHarmonies
-//---------------------------------------------------------
-
-static void alignHarmonies(const System* system, const std::vector<Segment*>& sl, bool harmony, const qreal maxShiftAbove,
-                           const qreal maxShiftBelow)
-{
-    // Help class.
-    // Contains harmonies/fretboard per segment.
-    class HarmonyList : public QList<Element*>
-    {
-        QMap<const Segment*, QList<Element*> > elements;
-        QList<Element*> modified;
-
-        Element* getReferenceElement(const Segment* s, bool above, bool visible) const
-        {
-            // Returns the reference element for aligning.
-            // When a segments contains multiple harmonies/fretboard, the lowest placed
-            // element (for placement above, otherwise the highest placed element) is
-            // used for alignment.
-            Element* element { nullptr };
-            for (Element* e : elements[s]) {
-                // Only chord symbols have styled offset, fretboards don't.
-                if (!e->autoplace() || (e->isHarmony() && !e->isStyled(Pid::OFFSET)) || (visible && !e->visible())) {
-                    continue;
-                }
-                if (!element) {
-                    element = e;
-                } else {
-                    if ((e->placeAbove() && above && (element->y() < e->y()))
-                        || (e->placeBelow() && !above && (element->y() > e->y()))) {
-                        element = e;
-                    }
-                }
-            }
-            return element;
-        }
-
-    public:
-        HarmonyList()
-        {
-            elements.clear();
-            modified.clear();
-        }
-
-        void append(const Segment* s, Element* e)
-        {
-            elements[s].append(e);
-        }
-
-        qreal getReferenceHeight(bool above) const
-        {
-            // The reference height is the height of
-            //    the lowest element if placed above
-            // or
-            //    the highest element if placed below.
-            bool first { true };
-            qreal ref { 0.0 };
-            for (auto s : elements.keys()) {
-                Element* e { getReferenceElement(s, above, true) };
-                if (!e) {
-                    continue;
-                }
-                if (e->placeAbove() && above) {
-                    ref = first ? e->y() : qMin(ref, e->y());
-                    first = false;
-                } else if (e->placeBelow() && !above) {
-                    ref = first ? e->y() : qMax(ref, e->y());
-                    first = false;
-                }
-            }
-            return ref;
-        }
-
-        bool align(bool above, qreal reference, qreal maxShift)
-        {
-            // Align the elements. If a segment contains multiple elements,
-            // only the reference elements is used in the algorithm. All other
-            // elements will remain their original placement with respect to
-            // the reference element.
-            bool moved { false };
-            if (almostZero(reference)) {
-                return moved;
-            }
-
-            for (auto s : elements.keys()) {
-                QList<Element*> handled;
-                Element* be = getReferenceElement(s, above, false);
-                if (!be) {
-                    // If there are only invisible elements, we have to use an invisible
-                    // element for alignment reference.
-                    be = getReferenceElement(s, above, true);
-                }
-                if (be && ((above && (be->y() < (reference + maxShift))) || ((!above && (be->y() > (reference - maxShift)))))) {
-                    qreal shift = be->rypos();
-                    be->rypos() = reference - be->ryoffset();
-                    shift -= be->rypos();
-                    for (Element* e : elements[s]) {
-                        if ((above && e->placeBelow()) || (!above && e->placeAbove())) {
-                            continue;
-                        }
-                        modified.append(e);
-                        handled.append(e);
-                        moved = true;
-                        if (e != be) {
-                            e->rypos() -= shift;
-                        }
-                    }
-                    for (auto e : handled) {
-                        elements[s].removeOne(e);
-                    }
-                }
-            }
-            return moved;
-        }
-
-        void addToSkyline(const System* system)
-        {
-            for (Element* e : qAsConst(modified)) {
-                const Segment* s = toSegment(e->parent());
-                const MeasureBase* m = toMeasureBase(s->parent());
-                system->staff(e->staffIdx())->skyline().add(e->shape().translated(e->pos() + s->pos() + m->pos()));
-                if (e->isFretDiagram()) {
-                    FretDiagram* fd = toFretDiagram(e);
-                    Harmony* h = fd->harmony();
-                    if (h) {
-                        system->staff(e->staffIdx())->skyline().add(h->shape().translated(h->pos() + fd->pos() + s->pos() + m->pos()));
-                    } else {
-                        system->staff(e->staffIdx())->skyline().add(fd->shape().translated(fd->pos() + s->pos() + m->pos()));
-                    }
-                }
-            }
-        }
-    };
-
-    if (almostZero(maxShiftAbove) && almostZero(maxShiftBelow)) {
-        return;
-    }
-
-    // Collect all fret diagrams and chord symbol and store them per staff.
-    // In the same pass, the maximum height is collected.
-    QMap<int, HarmonyList> staves;
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if ((harmony && e->isHarmony()) || (!harmony && e->isFretDiagram())) {
-                staves[e->staffIdx()].append(s, e);
-            }
-        }
-    }
-
-    for (int idx: staves.keys()) {
-        // Align the objects.
-        // Algorithm:
-        //    - Find highest placed harmony/fretdiagram.
-        //    - Align all harmony/fretdiagram objects placed between height and height-maxShiftAbove.
-        //    - Repeat for all harmony/fretdiagram objects below heigt-maxShiftAbove.
-        bool moved { true };
-        int pass { 0 };
-        while (moved && (pass++ < 10)) {
-            moved = false;
-            moved |= staves[idx].align(true, staves[idx].getReferenceHeight(true), maxShiftAbove);
-            moved |= staves[idx].align(false, staves[idx].getReferenceHeight(false), maxShiftBelow);
-        }
-
-        // Add all aligned objects to the sky line.
-        staves[idx].addToSkyline(system);
-    }
 }
