--- src/engraving/libmscore/layoutlinear.cpp
@@ -1,54 +1,0 @@
-/*
- * SPDX-License-Identifier: GPL-3.0-only
- * MuseScore-CLA-applies
- *
- * MuseScore
- * Music Composition & Notation
- *
- * Copyright (C) 2021 MuseScore BVBA and others
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as
- * published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-#include "score.h"
-#include "page.h"
-#include "system.h"
-#include "tremolo.h"
-#include "measure.h"
-#include "layout.h"
-#include "bracket.h"
-#include "spanner.h"
-#include "barline.h"
-#include "tie.h"
-#include "chord.h"
-#include "staff.h"
-#include "box.h"
-#include "spacer.h"
-#include "systemdivider.h"
-#include "tuplet.h"
-#include "dynamic.h"
-#include "stafflines.h"
-#include "tempotext.h"
-#include "hairpin.h"
-#include "part.h"
-#include "keysig.h"
-#include "sig.h"
-#include "breath.h"
-#include "tempo.h"
-#include "fermata.h"
-#include "lyrics.h"
-
-using namespace mu;
-
-namespace Ms {
-} // namespace Ms
