+++ src/engraving/style/styledef.h
@@ -1467,7 +1467,6 @@ enum class VerticalAlignRange {
     SEGMENT, MEASURE, SYSTEM
 };
 
-
 //---------------------------------------------------------
 //   StyledProperty
 ///   \cond PLUGIN_API \private \endcond
